+++ src/convertSchema.ts
@@ -52,7 +52,11 @@ const flagsToSql = (flags?: string[]) =>
 const renderCommaIf = (cond: boolean) => (cond ? ',' : '')
 const padding = '  '
 
-const _convertSchema = (nodes: Node[], spacing = ''): string => {
+const _convertSchema = (
+  nodes: Node[],
+  strictMode = false,
+  spacing = ''
+): string => {
   let returnVal = ''
   while (true) {
     const node = nodes[0]
@@ -65,9 +69,9 @@ const _convertSchema = (nodes: Node[], spacing = ''): string => {
     if (node.isRoot) {
       return (
         'CREATE TABLE IF NOT EXISTS %s (\n' +
-        _convertSchema(nodes.slice(1), padding) +
+        _convertSchema(nodes.slice(1), strictMode, padding) +
         ')' +
-        (node.val.additionalProperties
+        (strictMode && node.val.additionalProperties
           ? " WITH (column_policy = 'dynamic')"
           : '')
       )
@@ -94,11 +98,15 @@ const _convertSchema = (nodes: Node[], spacing = ''): string => {
         node.val.bsonType === 'array'
           ? 'ARRAY'
           : `OBJECT(${
-              node.val.additionalProperties === false ? 'STRICT' : 'DYNAMIC'
+              strictMode
+                ? node.val.additionalProperties === false
+                  ? 'STRICT'
+                  : 'DYNAMIC'
+                : 'IGNORED'
             }) AS`
       returnVal +=
         `${spacing}${field}${sqlType} (\n` +
-        _convertSchema(childNodes, newSpacing) +
+        _convertSchema(childNodes, strictMode, newSpacing) +
         `${spacing})${comma}\n`
       nodes = nodes.slice(index + 1)
     }
@@ -187,6 +195,6 @@ export const convertSchema: ConvertSchema = (
   if (options.overrides) {
     handleOverrides(nodes, options.overrides)
   }
-  const sqlSchema = _convertSchema(nodes)
+  const sqlSchema = _convertSchema(nodes, options.strictMode)
   return util.format(sqlSchema, qualifiedName)
 }